requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.115.0
Flask==3.1.1
fonttools==4.56.0
frozenlist==1.7.0
func_timeout==4.3.5
geneticalgorithm==1.0.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.6.3
numpy==2.2.3
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pluggy==1.6.0
propcache==0.3.2
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.11.0
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.2
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
